app/src/main/java/com/footballapps/egifcb/footballapps/teams/TeamsPresenter.kt
package com.footballapps.egifcb.footballapps.teams

import com.footballapps.egifcb.footballapps.api.ApiRepository
import com.footballapps.egifcb.footballapps.api.TheSportDBApi
import com.footballapps.egifcb.footballapps.model.ResponseTeam
import com.footballapps.egifcb.footballapps.until.CoroutineContextProvider
import com.google.gson.Gson
import kotlinx.coroutines.experimental.async
import org.jetbrains.anko.coroutines.experimental.bg

class TeamsPresenter(private val view: TeamsView,
                     private val apiRepository: ApiRepository,
                     private val gson: Gson,
                     private val context: CoroutineContextProvider = CoroutineContextProvider()) {

    fun getEventList(nameLeague: String?) {
        view.showLoading()

        async(context.main) {
            val data = bg {
                gson.fromJson(apiRepository.doRequest(TheSportDBApi.getTeams(nameLeague)),
                        ResponseTeam::class.java)
            }
            view.showEventList(data.await().teams)
            view.hideLoading()
        }
    }
}